Начало разработки бека на Golang, добавлен роутер создания матча. Создан логгер, создан парсер конфига, добавлен 1 роут для создания матча. Пока что не протестирован :clown:

+++ internal/models/match_type.go
@@ -1,7 +1,7 @@
 package models
 
 type MatchType struct {
-	ID      uint      `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID      uint64    `gorm:"primaryKey;autoIncrement" json:"id"`
 	Name    string    `gorm:"type:varchar(45);unique" json:"name"`
 	Matches []Matches `gorm:"foreignKey:MatchTypeId;references:ID"`
 }
